refactor(scheduler): scheduler.py(1078줄)를 책임별 6개 모듈로 분할
BackupScheduler(스케줄 엔진+실행 큐 직렬화)만 scheduler.py에 남기고
백업/SQL 실행, 로그 기록, 보관 정책을 별도 모듈로 분리했다.

- schedule_config.py: ScheduleTaskType/ScheduleConfig/_ExecutionJob/_ResolvedConnection
- cron_parser.py: CronParser
- execution_log_writer.py: ExecutionLogWriter (백업 로그 기록/조회, 파일 레이아웃 불변)
- retention_policy.py: select_paths_for_retention 공용 보관 정책 선정 헬퍼
- backup_task_executor.py: BackupTaskExecutor (RustDumpExporter 백업 실행)
- sql_query_task_executor.py: SqlQueryTaskExecutor (멀티 쿼리 실행 + 결과 저장)

scheduler.py는 ScheduleConfig/CronParser/BackupScheduler/ScheduleTaskType를
그대로 재노출하고, BackupScheduler의 공개 API 및 테스트가 참조하는 private
메서드(_execute_backup/_execute_sql_query/_execute_single_query/_parse_sql_queries)와
속성(_active_schedule_ids/_lock 등)을 위임 형태로 유지해 하위 호환을 보존한다.
create_rust_db_connector 조회는 BackupScheduler._make_connector 래퍼를 경유하도록 하여
기존 monkeypatch("src.core.scheduler.create_rust_db_connector") 테스트가 그대로 반영된다.
